Split packet_coding_test to narrow down MSVC build error
